+++ app/src/App.tsx
@@ -4,7 +4,6 @@ import {Toaster} from './components';
 import {Spinner} from './components/Spinner/Spinner';
 import {AppRouter} from './router/AppRouter/AppRouter';
 import {store} from './store';
-import './api/interceptors';
 
 export const App = (): JSX.Element => {
     return (
